# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 企业内部练习与考试系统 API
  version: 1.0.0
  description: |
    MVP 完整接口规格。全局约定见 [00-API设计说明.md](./00-API设计说明.md)。
    分域行为见 01~07 Markdown 文档。
  contact:
    name: Exam System Team
servers:
  - url: /api/v1
    description: 单租户企业内网 API
...
security:
  - sessionAuth: []
paths:
...
  /employees/{id}:
    get:
      tags: [Organization]
      operationId: getEmployee
      summary: 员工详情
      x-fr-ids: [FR-AUTH-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    patch:
      tags: [Organization]
      operationId: updateEmployee
      summary: 更新员工
      x-fr-ids: [FR-AUTH-04]
      x-page-ids: [AD-03, AD-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                departmentPath:
                  type: string
                phone:
                  type: string
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
  /import/template:
    get:
      tags: [Import]
      operationId: downloadImportTemplate
      summary: 下载导题模板
      x-fr-ids: [FR-IMP-01]
      x-page-ids: [AD-08]
      parameters:
        - name: questionBankId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: xlsx 模板
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /import/tasks:
    get:
      tags: [Import]
      operationId: listImportTasks
      summary: 导题任务列表
      x-fr-ids: [FR-IMP-05]
      x-page-ids: [AD-08]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    post:
      tags: [Import]
      operationId: createImportTask
      summary: 上传 Excel 创建导题任务
      x-fr-ids: [FR-IMP-01, FR-IMP-02]
      x-page-ids: [AD-08]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, questionBankId]
              properties:
                file:
                  type: string
                  format: binary
                questionBankId:
                  type: string
      responses:
        '201':
          $ref: '#/components/responses/SuccessEmpty'
  /import/tasks/{id}:
    get:
      tags: [Import]
      operationId: getImportTask
      summary: 导题任务详情
      x-fr-ids: [FR-IMP-05]
      x-page-ids: [AD-08, AD-09]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /import/tasks/{id}/preview:
    get:
      tags: [Import]
      operationId: getImportPreview
      summary: 校验预览
      x-fr-ids: [FR-IMP-03]
      x-page-ids: [AD-09]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 预览与 confirmToken
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImportPreview'
  /import/tasks/{id}/confirm:
    post:
      tags: [Import]
      operationId: confirmImportTask
      summary: 幂等确认导入
      x-fr-ids: [FR-IMP-04]
      x-page-ids: [AD-09]
      x-sys-ids: [SYS-02]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmToken]
              properties:
                confirmToken:
                  type: string
                confirmPendingHierarchy:
                  type: boolean
                  default: true
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
        '409':
          $ref: '#/components/responses/Conflict'
  /import/tasks/{id}/cancel:
    post:
      tags: [Import]
      operationId: cancelImportTask
      summary: 取消导题任务
      x-fr-ids: [FR-IMP-05]
      x-page-ids: [AD-08]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /import/tasks/{id}/revalidate:
    post:
      tags: [Import]
      operationId: revalidateImportTask
      summary: 重新校验
      x-fr-ids: [FR-IMP-05]
      x-page-ids: [AD-09]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /import/tasks/{id}/errors:
    get:
      tags: [Import]
      operationId: downloadImportErrors
      summary: 下载错误报告
      x-fr-ids: [FR-IMP-02]
      x-page-ids: [AD-09]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 错误 xlsx
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
...
  /admin/exams:
    get:
      tags: [Exam]
      operationId: listAdminExams
      summary: 考试列表
      x-fr-ids: [FR-EXM-01]
      x-page-ids: [AD-10]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    post:
      tags: [Exam]
      operationId: createExam
      summary: 创建考试草稿
      x-fr-ids: [FR-EXM-01]
      x-page-ids: [AD-10, AD-11]
      responses:
        '201':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}:
    get:
      tags: [Exam]
      operationId: getAdminExam
      summary: 考试详情
      x-fr-ids: [FR-EXM-01]
      x-page-ids: [AD-11]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    patch:
      tags: [Exam]
      operationId: patchAdminExam
      summary: 更新考试（草稿或说明）
      x-fr-ids: [FR-EXM-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/wizard/basic:
    put:
      tags: [Exam]
      operationId: updateExamWizardBasic
      summary: 向导-基础信息
      x-fr-ids: [FR-EXM-01]
      x-page-ids: [AD-11]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/wizard/rules:
    put:
      tags: [Exam]
      operationId: updateExamWizardRules
      summary: 向导-规则与组卷
      x-fr-ids: [FR-EXM-01, FR-EXM-02]
      x-page-ids: [AD-11]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/wizard/assignees:
    put:
      tags: [Exam]
      operationId: updateExamWizardAssignees
      summary: 向导-应考人员
      x-fr-ids: [FR-EXM-01]
      x-page-ids: [AD-11]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/wizard/visibility:
    put:
      tags: [Exam]
      operationId: updateExamWizardVisibility
      summary: 向导-公开策略
      x-fr-ids: [FR-EXM-01, FR-SCR-04]
      x-page-ids: [AD-11]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/wizard/review:
    put:
      tags: [Exam]
      operationId: updateExamWizardReview
      summary: 向导-摘要确认
      x-fr-ids: [FR-EXM-01]
      x-page-ids: [AD-11]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/preflight:
    post:
      tags: [Exam]
      operationId: examPreflight
      summary: 发布预检
      x-fr-ids: [FR-EXM-02]
      x-page-ids: [AD-12]
      x-sys-ids: [SYS-03]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/publish:
    post:
      tags: [Exam]
      operationId: publishExam
      summary: 发布考试
      x-fr-ids: [FR-EXM-03]
      x-page-ids: [AD-12]
      x-sys-ids: [SYS-03]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/cancel:
    post:
      tags: [Exam]
      operationId: cancelExam
      summary: 取消考试
      x-fr-ids: [FR-EXM-04]
      x-page-ids: [AD-10, AD-13]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeVisibleReason, internalReason]
              properties:
                employeeVisibleReason:
                  type: string
                internalReason:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/monitor:
    get:
      tags: [Exam]
      operationId: getExamMonitor
      summary: 过程监控
      x-fr-ids: [FR-EXM-05, FR-REP-01]
      x-page-ids: [AD-13]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
  /exams/{id}:
    get:
      tags: [Exam]
      operationId: getExamTaskDetail
      summary: 任务详情（脱敏）
      x-fr-ids: [FR-EXM-05, FR-SCR-04]
      x-page-ids: [MP-10, EX-03]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /exams/records:
    get:
      tags: [Exam]
      operationId: listExamRecords
      summary: 正式考试学习记录
      x-fr-ids: [FR-EXM-05]
      x-page-ids: [MP-11]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 亦可用 HttpOnly Session Cookie，二者等价
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 幂等键，见 00-API设计说明
  responses:
    SuccessEmpty:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
...
    Conflict:
      description: 版本或状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required: [serverNow, requestId]
      properties:
        serverNow:
          type: string
          format: date-time
        requestId:
          type: string
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - AUTH_INVALID_CREDENTIALS
        - AUTH_ACCOUNT_LOCKED
        - AUTH_ACCOUNT_DISABLED
        - AUTH_SESSION_EXPIRED
        - AUTH_PASSWORD_CHANGE_REQUIRED
        - AUTH_SMS_RATE_LIMITED
        - ORG_DEPARTMENT_NAME_DUPLICATE
        - ORG_LAST_OUTAGE_ADMIN
        - IMP_PREVIEW_STALE
        - PRA_SESSION_ALREADY_ACTIVE
        - SIM_ATTEMPT_ALREADY_ACTIVE
        - ATT_NO_REMAINING_OPPORTUNITY
        - ATT_EXAM_PAUSED
        - ANS_VERSION_CONFLICT
        - ANS_UNCONFIRMED_ANSWERS
        - ANS_ATTEMPT_TERMINATED
        - SCR_SUBMIT_IN_PROGRESS
        - OPS_PROPOSAL_STALE
        - OPS_NOT_AUTHORIZED
        - SEC_FORBIDDEN
...
    ImportPreview:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
        confirmToken:
          type: string
        importableCount:
          type: integer
        errorCount:
          type: integer
        pendingHierarchy:
          type: object
        previewExpiresAt:
          type: string
          format: date-time
...